server/requirements.txt
fastapi
uvicorn[standard]
gunicorn

sqlalchemy
psycopg[binary]

pydantic
python-jose[cryptography]

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-multipart
aiofiles

sqladmin
jinja2
email-validator

itsdangerous>=2.1